Updates to the cognito post auth trigger
