+++ source/cognito-post-confirmation-trigger/lib/error.js
