--- source/cognito-triggers/post-confirmation/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "CommonJS",
-    "target": "ES2017",
-    "noImplicitAny": true,
-    "preserveConstEnums": true,
-    "outDir": "./built",
-    "sourceMap": true
-  },
-  "include": ["index.ts"],
-  "exclude": ["node_modules", "**/*.spec.ts"]
-}
